+++ packages/build-mpa-config/src/index.ts
@@ -1,6 +1,7 @@
 import * as path from 'path';
 import { formatPath, checkExportDefaultDeclarationExists } from '@builder/app-helpers';
-import generateEntry from './generateEntry';
+import generateEntry from './generate';
+import { FrameworkType, IGenerateResult } from './types';
 
 interface IEntries {
   entryName: string;
@@ -11,7 +12,7 @@ interface IEntries {
 
 interface IConfigOptions {
   type?: string;
-  framework?: string;
+  framework?: FrameworkType;
   entries?: IEntries[];
   targetDir?: string;
 }
@@ -31,45 +32,68 @@ export const generateMPAEntries = (api, options: IConfigOptions) => {
   entries.forEach((entry) => {
     const { entryName, entryPath, ...pageConfig } = entry;
     const { source } = pageConfig;
-    const useOriginEntry = /app(\.(t|j)sx?)?$/.test(entryPath);
-    // icejs will config entry by api modifyUserConfig
-    // when the entry has no export default declaration or is app.ts, do not generate entry
-    const finalEntry = !useOriginEntry && checkExportDefaultDeclarationExists(path.join(rootDir, 'src', source)) ?
-      generateEntry(api, { framework, targetDir, pageEntry: entryPath, entryName, pageConfig }) :
-      entryPath;
+    const isAppEntry = /app(\.(t|j)sx?)?$/.test(entryPath);
+    // when the entry has no export default declaration, do not generate any files
+    let finalEntry = entryPath;
+    let runAppPath = null;
+    if (isAppEntry || checkExportDefaultDeclarationExists(path.join(rootDir, 'src', source))) {
+      const result = generateEntry(api, { framework, targetDir, pageEntry: entryPath, entryName, pageConfig, isAppEntry });
+      finalEntry = result.entryPath;
+      runAppPath = result.runAppPath;
+    }
 
     parsedEntries[entryName] = {
       ...entry,
       finalEntry,
+      shouldRedirectRunApp: isAppEntry,
+      runAppPath,
     };
   });
   return parsedEntries;
+};
+
+export const addRedirectRunAppLoader = (config: any, framework, redirectEntries: IGenerateResult[] = []) => {
+  // TODO: esbuild preCompile ts to js
+  config.module.rule('redirect-runApp')
+    .enforce('post')
+    .test(filepath => redirectEntries.some(({ entryPath }) => entryPath === filepath))
+    .use('redirect-runApp-loader')
+    .loader(require.resolve(path.join(__dirname, 'redirectRunAppLoader')))
+    .options({
+      framework,
+      redirectEntries,
+    });
 };
 
 const setMPAConfig = (api, config, options: IConfigOptions) => {
   if (!options) {
     throw new Error('There need pass options param to setMPAConfig method');
   }
-  const { type = 'web' } = options;
   const parsedEntries = generateMPAEntries(api, options);
 
-  // do not splitChunks when mpa
-  config.optimization.splitChunks({ cacheGroups: {} });
   // clear entry points
   config.entryPoints.clear();
   // add mpa entries
   const matchStrs = [];
-
+  // add redirect entry path
+  const redirectEntries: IGenerateResult[] = [];
   Object.keys(parsedEntries).forEach((entryKey) => {
-    const { entryName, source, finalEntry } = parsedEntries[entryKey];
+    const { entryName, source, finalEntry, shouldRedirectRunApp, runAppPath } = parsedEntries[entryKey];
     config.entry(entryName).add(finalEntry);
-
+    if (shouldRedirectRunApp) {
+      redirectEntries.push({
+        entryPath: finalEntry,
+        runAppPath,
+      });
+    }
     // get page paths for rule match
     const matchStr = `src/${source}`;
     matchStrs.push(formatPath(matchStr));
   });
 
-  if (type === 'web' && config.plugins.has('document')) {
+  addRedirectRunAppLoader(config, redirectEntries);
+
+  if (config.plugins.has('document')) {
     config.plugin('document').tap(args => {
       return [{
         ...args[0],
@@ -79,13 +103,11 @@ const setMPAConfig = (api, config, options: IConfigOptions) => {
   }
 
   // modify appJSON rules for mpa
-  if (config.module.rules.get('appJSON')) {
-    const matchInclude = (filepath: string) => {
-      const matchReg = matchStrs.length ? new RegExp(matchStrs.join('|')) : null;
-      return matchReg && matchReg.test(filepath);
-    };
-    config.module.rule('appJSON').include.add(matchInclude);
-  }
+  const matchInclude = (filepath: string) => {
+    const matchReg = matchStrs.length ? new RegExp(matchStrs.join('|')) : null;
+    return matchReg && matchReg.test(filepath);
+  };
+  config.module.rule('appJSON').include.add(matchInclude);
 };
 
 export default setMPAConfig;