+++ packages/plugin-request/template/request.ts
@@ -1,20 +1,28 @@
-import { AxiosRequestConfig } from 'axios';
+import axios, { AxiosRequestConfig, CancelTokenStatic, CancelStatic } from 'axios';
 import * as utils from 'axios/lib/utils';
 import createAxiosInstance from './createAxiosInstance';
 
+interface IRequestConfig extends AxiosRequestConfig {
+  instanceName?: string;
+  withFullResponse?: boolean;
+}
+
 export interface IRequestProps {
-  get: <T = any>(url: string, config?: AxiosRequestConfig) => Promise<T>;
-  delete: <T = any>(url: string, config?: AxiosRequestConfig) => Promise<T>;
-  head: <T = any>(url: string, config?: AxiosRequestConfig) => Promise<T>;
-  options: <T = any>(url: string, config?: AxiosRequestConfig) => Promise<T>;
-  post: <T = any>(url: string, data?: any, config?: AxiosRequestConfig) => Promise<T>;
-  put: <T = any>(url: string, data?: any, config?: AxiosRequestConfig) => Promise<T>;
-  patch: <T = any>(url: string, data?: any, config?: AxiosRequestConfig) => Promise<T>;
+  get: <T = any>(url: string, config?: IRequestConfig) => Promise<T>;
+  delete: <T = any>(url: string, config?: IRequestConfig) => Promise<T>;
+  head: <T = any>(url: string, config?: IRequestConfig) => Promise<T>;
+  options: <T = any>(url: string, config?: IRequestConfig) => Promise<T>;
+  post: <T = any>(url: string, data?: any, config?: IRequestConfig) => Promise<T>;
+  put: <T = any>(url: string, data?: any, config?: IRequestConfig) => Promise<T>;
+  patch: <T = any>(url: string, data?: any, config?: IRequestConfig) => Promise<T>;
 }
 
 interface IRequest extends IRequestProps {
-  <T = any>(options: AxiosRequestConfig): Promise<T>;
-  <T = any>(url: string, config?: AxiosRequestConfig): Promise<T>;
+  <T = any>(options: IRequestConfig): Promise<T>;
+  <T = any>(url: string, config?: IRequestConfig): Promise<T>;
+  Cancel: CancelStatic;
+  CancelToken: CancelTokenStatic;
+  isCancel(value: any): boolean;
 }
 
 /**
@@ -59,5 +67,8 @@ utils.forEach(['post', 'put', 'patch'], function forEachMethodWithData(method) {
     }));
   };
 });
+
+request.CancelToken = axios.CancelToken;
+request.isCancel = axios.isCancel;
 
 export default request as IRequest;