+++ packages/plugin-request/CHANGELOG.md
@@ -4,6 +4,14 @@
 
 - [feat] migrate plugin api and refactor runtime folder
 
+## 1.7.8
+
+- [feat] provide types lint
+
+## 1.7.7
+
+- [feat] support axios cancel
+
 ## 1.7.6
 
 - [chore] bump version